Moved Person class to seperate file
